[M146] metrics: Track Incognito NTP tab height change on omnibox focus.
Original change's description:
> metrics: Track Incognito NTP tab height change on omnibox focus.
> 
> This change introduces a new boolean histogram,
> "NewTabPage.Incognito.OmniboxAutofocus.OnFocus.TabHeightChange", to
> record whether the Incognito NTP tab's non-zero height decreased after
> the omnibox was focused.
> 
> The other layout metrics related to keyboard visibility are now only
> recorded if the tab height was indeed decreased to non-zero value, which
> is expected when the on-screen keyboard appears.
> 
> This helps to filter out potentially inaccurate measurements and
> provides better insight into the conditions under which the layout
> metrics are collected.
> 
> Unit tests are added to verify the behavior of recording histograms
> based on whether the tab height is decreased.
> 
> Fixed: 481976427, 481979913
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7548726
> Commit-Queue: Matt Dembski <dembski@google.com>
> Cr-Commit-Position: refs/heads/main@{#1585519}

(cherry picked from commit bdc84494efaf806b7ccbdaa9b1fcceabc00df0c8)

Bug: 485472746,481976427,481979913
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7588933
Bot-Commit: Rubber Stamper <rubber-stamper@appspot.gserviceaccount.com>
Auto-Submit: Chrome Cherry Picker <chrome-cherry-picker@chops-service-accounts.iam.gserviceaccount.com>
Commit-Queue: Rubber Stamper <rubber-stamper@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/branch-heads/7680@{#616}
Cr-Branched-From: 76b7d80e5cda23fe6537eed26d68c92e995c7f39-refs/heads/main@{#1582197}

+++ chrome/android/java/src/org/chromium/chrome/browser/ntp/IncognitoNtpOmniboxAutofocusManager.java
@@ -88,7 +88,7 @@ public class IncognitoNtpOmniboxAutofocusManager {
     private int mNtpOpenedCount;
     private final @NonNull GestureDetector mNtpSingleTapDetector;
     private boolean mIsAutofocusing;
-    private double mTabHeightBeforeFocus;
+    private int mTabHeightBeforeFocus;
 
     /**
      * Overrides the result of {@link #checkAutofocusAllowedWithPrediction(Tab)} for testing
